perf: 렌더링 성능 최적화를 위한 번들 및 리소스 최적화
- React Query 조건부 로딩: 루트 레이아웃에서 제거하고 검색 페이지만 사용
  - 초기 JavaScript 번들 크기 약 30-40KB 감소 (TTB 개선)

- 로컬 폰트 최적화: next/font/local로 전환하여 자동 프리로드 및 최적화
  - D2Coding, FZuanSu 폰트를 next/font/local로 마이그레이션
  - FCP 100-200ms 개선 예상

- PrismJS CSS 조건부 로드: 코드 블록이 있는 레코드 페이지만 로드
  - 초기 CSS 크기 약 5-10KB 감소

- 검색 페이지 리팩토링: 서버 컴포넌트와 클라이언트 컴포넌트 분리
  - SearchContent를 별도 클라이언트 컴포넌트로 분리
  - QueryProvider를 검색 페이지만 감싸도록 변경

- 루트 페이지 최적화: Suspense 제거로 완전한 SSG 적용

변경된 파일:
- src/app/layout.tsx: QueryProvider 제거, 로컬 폰트 추가
- src/app/page.tsx: Suspense 제거
- src/app/search/page.tsx: 서버 컴포넌트로 변경, SearchContent 분리
- src/app/search/SearchContent.tsx: 새 파일 (검색 로직 클라이언트 컴포넌트)
- src/app/styles/globals.css: PrismJS CSS import 제거
- src/app/styles/theme.css: 폰트 변수명 업데이트
- src/app/(record)/records/[slug]/layout.tsx: PrismJS CSS 추가
