+++ src/app/styles/globals.css
@@ -1,8 +1,6 @@
 @import "tailwindcss";
-@import "./font.css";
 @import "./theme.css" layer(theme);
 @import "./utility.css";
-@import "prismjs/themes/prism-tomorrow.css";
 
 @layer base {
   :root {
